requirements.txt
# Core Dependencies
streamlit      # Streamlit for UI
pymupdf          # PyMuPDF (fitz) for PDF processing
pandas              # DataFrame support for Excel and CSV files
python-pptx       # PowerPoint file processing
pytesseract      # OCR engine integration with Tesseract
pillow         # Image processing library (PIL)
llama-index-core   # Core LlamaIndex library for document indexing
llama-index-llms-groq  # Groq LLM integration with LlamaIndex
llama-index-embeddings-huggingface  # HuggingFace embeddings integration
groq             # Groq LLM API library

# General Document Processing Libraries
textract        # General-purpose text extraction from files

# Image Preprocessing Libraries
opencv-python      # OpenCV for image preprocessing (grayscale, thresholding, etc.)
numpy          # Numerical operations on images

# Progress Bar Library (Optional)
tqdm            # Progress bar for long-running tasks

# Hugging Face Models and Transformers
sentence-transformers  # Sentence embeddings from Hugging Face

# Optional Dependencies (for specific file types)
xlrd             # Legacy Excel file support (.xls)
python-docx        # Word document (.docx) processing
odfpy         # OpenDocument format support (.odt)

